+++ M2TWEOP Code/M2TWEOP library/smallFuncs.h
@@ -27,7 +27,7 @@ namespace smallFuncs
 	//toggle the highlighting of units on the tactical map.
 	NOINLINE EOP_EXPORT void swUnBMapHighlight();
 	// Zoom out the camera beyond the normal limit
-	NOINLINE EOP_EXPORT void changeBattlemapCameraDistance(signed short maxZoom);
+	NOINLINE EOP_EXPORT void changeBattlemapCameraDistance(int maxZoom);
 
 	NOINLINE EOP_EXPORT void setReligionsLimit(unsigned char limit);
